+++ README.md
@@ -1,16 +1,2 @@
-<h1>Test Pro  Class Automation Framework Repository</h1><br>
-
-Welcome to the Test Pro A53 Class Automation Framework repository! <br>
-
-<h2>Overview</h2>
-This repository contains the Test Pro A53 Class Automation Framework, which serves as an educational tool for students' homework and live class demonstrations.
-
-<h2>Getting Started</h2>
-If you're a student enrolled in the A53 class, follow these steps to get started with the A53 Class Automation Framework:<br><br>
-1.) Clone this repository to your local machine. <br>
-2.) Explore the sample test case in the 'LoginTests' java class located in the 'src' directory under 'test' > 'java' folder <br>
-3.) Complete the assigned homework exercises and participate in live class demonstrations using the framework. <br><br>
-For more detailed instructions regarding the installation of the framework, please refer to the A53 class materials (LMS & slides).
-
-<h2>Contact Information</h2>
-If you have any questions or need further assistance, feel free to contact the course instructor, our support team or send a message in Teams for A53.  
+# musicPlayer
+ This is a testNG framework using java to automate a music player features
