+++ build.gradle
@@ -16,7 +16,13 @@ dependencies {
 }
 
 test {
-    useTestNG() 
+    useTestNG() {
+        //suites "src/test/resources/Regression.xml"
+        //suites "src/test/resources/Smoke.xml"
+     //   suites "src/test/resources/${suite}.xml"
+     //   systemProperty "browser", System.getProperty("browser")
+      //  systemProperty "suite", System.getProperty("suite")
+    }
 }
 
 /*task testSmoke(type: Test){
