+++ src/test/java/BaseTest.java
@@ -1,0 +1,33 @@
+import io.github.bonigarcia.wdm.WebDriverManager;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
+import org.openqa.selenium.support.ui.WebDriverWait;
+import org.testng.annotations.AfterTest;
+import org.testng.annotations.BeforeClass;
+
+import java.time.Duration;
+
+public class BaseTest {
+    public WebDriver driver;
+    public WebDriverWait wait;
+
+    @BeforeClass
+    public void setUpDriver() {
+        WebDriverManager.chromedriver().setup();
+        ChromeOptions options = new ChromeOptions();
+        options.addArguments("--disable-notifications", "--remote-allow-origins=*", "--incognito", "--start-maximized");
+        options.setExperimentalOption("excludeSwitches", new String[]{"enable-automation"});
+        driver = new ChromeDriver(options);
+        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
+        driver.manage().window().maximize();
+        String url = "https://qa.koel.app"; // Overriding URL with specified value
+        driver.get(url); // Navigating WebDriver to the specified URL
+        wait = new WebDriverWait(driver, Duration.ofSeconds(10));
+    }
+
+    @AfterTest
+    public void closeBrowser() {
+        driver.quit();
+    }
+}
