+++ src/test/resources/TestNG.xml
@@ -7,13 +7,7 @@
 
     <test name="Example Test" preserve-order="false">
         <classes>
-            <class name="Pages.Homework15"/>
-            <!--<class name="Pages.BaseTest"/>
-            <class name="HomeWork18"/>
-            <class name="Homework19"/>
-            <class name="AllSongsTest"/>
-            <class name="Homework22"/>
-            <class name="ProfileTest"/>-->
+            <class name="Homework_15"/>
         </classes>
     </test>
 </suite>
